u32 to u16 fix build

+++ boards/BMS/Test/Src/Test_Imd.cpp
@@ -2,7 +2,7 @@
 
 FAKE_VALUE_FUNC(float, get_pwm_frequency);
 FAKE_VALUE_FUNC(float, get_pwm_duty_cycle);
-FAKE_VALUE_FUNC(uint32_t, get_seconds_since_power_on);
+FAKE_VALUE_FUNC(uint16_t, get_seconds_since_power_on);
 
 void ImdTest::SetImdCondition(
     struct Imd *       imd_to_set,
